Rewrites Transfer with only compressed_history (insteadt of prev_...)
