+++ nnfabrik/templates/__init__.py
@@ -1,3 +1,3 @@
 from .trained_model import TrainedModelBase, DataInfoBase
 from .scoring import ScoringBase
-from .transfer import TransferredTrainedModelBase
+from .transfer.transferred_trained_model import TransferredTrainedModelBase
